Revert "Package structure non-breaking climate fixes"
