+++ src/main/java/ru/mail/polis/lsm/artem_drozdov/LsmDAO.java
@@ -3,7 +3,6 @@ package ru.mail.polis.lsm.artem_drozdov;
 import ru.mail.polis.lsm.DAO;
 import ru.mail.polis.lsm.DAOConfig;
 import ru.mail.polis.lsm.Record;
-import ru.mail.polis.lsm.artem_drozdov.util.RecordIterator;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.GuardedBy;
@@ -12,9 +11,12 @@ import java.io.UncheckedIOException;
 import java.nio.ByteBuffer;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NavigableMap;
+import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.SortedMap;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.ConcurrentSkipListMap;
@@ -41,11 +43,11 @@ public class LsmDAO implements DAO {
         synchronized (this) {
             Iterator<Record> sstableRanges = sstableRanges(fromKey, toKey);
             Iterator<Record> memoryRange = map(fromKey, toKey).values().iterator();
-            Iterator<Record> iterator = RecordIterator.mergeTwo(
-                    new RecordIterator.PeekingIterator(sstableRanges),
-                    new RecordIterator.PeekingIterator(memoryRange)
+            Iterator<Record> iterator = mergeTwo(
+                    new PeekingIterator(sstableRanges),
+                    new PeekingIterator(memoryRange)
             );
-            return RecordIterator.filterTombstones(iterator);
+            return filterTombstones(iterator);
         }
     }
 
@@ -111,7 +113,7 @@ public class LsmDAO implements DAO {
         for (SSTable ssTable : tables) {
             iterators.add(ssTable.range(fromKey, toKey));
         }
-        return RecordIterator.merge(iterators);
+        return merge(iterators);
     }
 
     private SortedMap<ByteBuffer, Record> map(@Nullable ByteBuffer fromKey, @Nullable ByteBuffer toKey) {
@@ -125,5 +127,129 @@ public class LsmDAO implements DAO {
             return memoryStorage.tailMap(fromKey);
         }
         return memoryStorage.subMap(fromKey, toKey);
+    }
+
+    private static Iterator<Record> merge(List<Iterator<Record>> iterators) {
+        if (iterators.isEmpty()) {
+            return Collections.emptyIterator();
+        }
+        if (iterators.size() == 1) {
+            return iterators.get(0);
+        }
+        if (iterators.size() == 2) {
+            return mergeTwo(new PeekingIterator(iterators.get(0)), new PeekingIterator(iterators.get(1)));
+        }
+        Iterator<Record> left = merge(iterators.subList(0, iterators.size() / 2));
+        Iterator<Record> right = merge(iterators.subList(iterators.size() / 2, iterators.size()));
+        return mergeTwo(new PeekingIterator(left), new PeekingIterator(right));
+    }
+
+    private static Iterator<Record> mergeTwo(PeekingIterator left, PeekingIterator right) {
+        return new Iterator<>() {
+
+            @Override
+            public boolean hasNext() {
+                return left.hasNext() || right.hasNext();
+            }
+
+            @Override
+            public Record next() {
+                if (!hasNext()) {
+                    throw new NoSuchElementException("No elements");
+                }
+
+                if (!left.hasNext()) {
+                    return right.next();
+                }
+                if (!right.hasNext()) {
+                    return left.next();
+                }
+
+                // checked earlier
+                ByteBuffer leftKey = Objects.requireNonNull(left.peek()).getKey();
+                ByteBuffer rightKey = Objects.requireNonNull(right.peek()).getKey();
+
+                int compareResult = leftKey.compareTo(rightKey);
+                if (compareResult == 0) {
+                    left.next();
+                    return right.next();
+                }
+
+                if (compareResult < 0) {
+                    return left.next();
+                } else {
+                    return right.next();
+                }
+            }
+
+        };
+    }
+
+    private static Iterator<Record> filterTombstones(Iterator<Record> iterator) {
+        PeekingIterator delegate = new PeekingIterator(iterator);
+        return new Iterator<>() {
+            @Override
+            public boolean hasNext() {
+                for (;;) {
+                    Record peek = delegate.peek();
+                    if (peek == null) {
+                        return false;
+                    }
+                    if (!peek.isTombstone()) {
+                        return true;
+                    }
+
+                    delegate.next();
+                }
+            }
+
+            @Override
+            public Record next() {
+                if (!hasNext()) {
+                    throw new NoSuchElementException("No elements");
+                }
+                return delegate.next();
+            }
+        };
+    }
+
+    private static class PeekingIterator implements Iterator<Record> {
+
+        private Record current;
+
+        private final Iterator<Record> delegate;
+
+        public PeekingIterator(Iterator<Record> delegate) {
+            this.delegate = delegate;
+        }
+
+        @Override
+        public boolean hasNext() {
+            return current != null || delegate.hasNext();
+        }
+
+        @Override
+        public Record next() {
+            if (!hasNext()) {
+                throw new NoSuchElementException();
+            }
+            Record now = peek();
+            current = null;
+            return now;
+        }
+
+        public Record peek() {
+            if (current != null) {
+                return current;
+            }
+
+            if (!delegate.hasNext()) {
+                return null;
+            }
+
+            current = delegate.next();
+            return current;
+        }
+
     }
 }
